% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular-math.R
\name{azimuth_difference}
\alias{azimuth_difference}
\title{Compute a a difference between two azimuth using angular math}
\usage{
azimuth_difference(azi_1, azi_2)
}
\arguments{
\item{azi_1}{First azimuth in decimal degrees}

\item{azi_2}{Second azimuth in decimal degrees}
}
\value{
Difference in decimal degrees
}
\description{
See this slide deck where Bryan pulled this formula from: http://webspace.ship.edu/pgmarr/Geo441/Lectures/Lec\%2016\%20-\%20Directional\%20Statistics.pdf
}
\examples{
azimuth_difference(355,10)
}
