% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{get_sharepoint_dir}
\alias{get_sharepoint_dir}
\title{Get sharepoint directory for user running this function}
\usage{
get_sharepoint_dir()
}
\value{
Your sharepoint directory for appending further paths to
}
\description{
Get sharepoint directory for user running this function
}
