% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular-math.R
\name{mean_azimuth}
\alias{mean_azimuth}
\title{Compute a weighted average azimuth (using angular statistics) of multiple line segments}
\usage{
mean_azimuth(azi, len)
}
\arguments{
\item{azi}{Azimuth of line in decimal degrees}

\item{len}{Length of line in planar unit (e.g., feet)}
}
\value{
Weighted average azmimuth of line segments in decimal degrees
}
\description{
See this slide deck where Bryan pulled this formula from: http://webspace.ship.edu/pgmarr/Geo441/Lectures/Lec\%2016\%20-\%20Directional\%20Statistics.pdf
}
\examples{
mean_azimuth(c(355,5),c(10,5))
}
