% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{mm_norm}
\alias{mm_norm}
\title{Min-Max Normalization of a Vector of Numbers}
\usage{
mm_norm(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Normalized vector
}
\description{
I have typically used this when evaluating between choosing a particular reference segment on the basis of multiple metrics, e.g. distance, bearing similarity, length overlap.
}
\examples{
mm_norm(c(1,2,3,4,5,6,7,8,9,10))
}
