% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nn_base_color_palette}
\alias{nn_base_color_palette}
\title{Base NN Color Palette (2022)}
\format{
A data frame with 18 rows and 3 variables
\describe{
\item{color_name}{NN Name of Color}
\item{hex_code}{Hexadecimal code definition of color}
\item{color_category}{Recommended category of color (e.g., primary, secondary)}
}
}
\source{
Brendan Rahman, February 1st, 2022
}
\usage{
nn_base_color_palette
}
\description{
Definition (hex codes and categories) for 2022 NN colors
}
\keyword{datasets}
