% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{nn_basic_theme}
\alias{nn_basic_theme}
\title{Basic Nelson\\Nygaard Plot Theme}
\usage{
nn_basic_theme(
  legend_right = FALSE,
  grey_background = TRUE,
  base_size = 12,
  base_family = "Barlow",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\arguments{
\item{legend_right}{Legend defaults to top, if you want on right, set this parameter}

\item{grey_background}{By default, these plots have a light grey background. Set to FALSE to use white.}

\item{base_size}{Base font size, defaults to 12}

\item{base_family}{Base font family, defaults to Barlow}

\item{base_line_size}{Base grid line size}

\item{base_rect_size}{Base rectangle size}
}
\description{
Basic Nelson\\Nygaard Plot Theme
}
