% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{nn_color_ramp}
\alias{nn_color_ramp}
\title{Return a sequential, diverging, or qualitative palette/ramp of NN colors}
\usage{
nn_color_ramp(
  num_colors = 5,
  type = "seq",
  palette_name = "Blue",
  direction = 1
)
}
\arguments{
\item{num_colors}{Number of colors you would like, within the bounds of # of colors available}

\item{type}{One of seq (sequential), div (diverging) or qual (qualitative)}

\item{palette_name}{String indicating the name of the palette (not case-sensitive)}

\item{direction}{Sets the order of colors in the scale. If 1, colors are ordered in the default order. If -1, the order of colors is reversed.}
}
\value{
Character vector of hex codes corresponding to queried colors
}
\description{
Return a sequential, diverging, or qualitative palette/ramp of NN colors
}
\examples{
nn_color_ramp(num_colors = 5, type = "seq", palette_name = "Blue", direction=1)
}
