% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{nn_colors}
\alias{nn_colors}
\title{Return NN colors}
\usage{
nn_colors(...)
}
\arguments{
\item{...}{Can leave empty if you just want to return all colors, specify name(s) if you want to return those colors}
}
\value{
All colors, or requested colors (hex codes)
}
\description{
Return NN colors
}
\examples{
nn_colors()
nn_colors("NN Blue")
}
