% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular-math.R
\name{opposite_azimuth}
\alias{opposite_azimuth}
\title{Return opposite azimuth in decimal degrees (for a given azimuth in decimal degrees)}
\usage{
opposite_azimuth(azi)
}
\arguments{
\item{azi}{Input azimuth (decimal degrees)}
}
\value{
Opposite azimuth (decimal degrees)
}
\description{
Return opposite azimuth in decimal degrees (for a given azimuth in decimal degrees)
}
\examples{
opposite_azimuth(270)
}
