% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{replace_nan}
\alias{replace_nan}
\title{Replance NaN with NA}
\usage{
replace_nan(x)
}
\arguments{
\item{x}{Vector (can be used for a data frame column)}
}
\value{
Vector with NaNs replaced with NAs
}
\description{
Replance NaN with NA
}
\examples{
replace_nan(c(0, 1, 2, NaN))
}
