% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular-math.R
\name{st_avg_line_azimuth}
\alias{st_avg_line_azimuth}
\title{Compute the average azimuth of a complex (typically more than two vertices) line segment}
\usage{
st_avg_line_azimuth(linestring, coord_local)
}
\arguments{
\item{linestring}{Input linestring (from {sf})}

\item{coord_local}{Local coordinate system EPSG code of input linestring}
}
\value{
Average azimuth in decimal degrees
}
\description{
Computes the azimuth and length for every sub-segment (i.e. the line between two successive vertices) and then takes a length-weighted average of all sub-segments.
Typically used within a {purrr} \code{map} function.
}
