% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset-lines.R
\name{st_parallel_line}
\alias{st_parallel_line}
\title{Create parallel lines at distance, assumes sf data frame input.}
\usage{
st_parallel_line(line_obj, coord_local, distance, forward = TRUE)
}
\arguments{
\item{line_obj}{\code{{sf}} data frame}

\item{coord_local}{EPSG code of local (planar) coordinate system}

\item{distance}{Distance of parallel line, in specified planar coordinate system}

\item{forward}{Direction of parallel line generation. Default is in forward direction, select FALSE if you want in other direction.}
}
\value{
\code{{sf}} data frame with transformed geometry (parallel lines per specifications)
}
\description{
Create parallel lines at distance, assumes sf data frame input.
}
