% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-clip-custom.R
\name{write_clip_with_header}
\alias{write_clip_with_header}
\title{Writes to Clipboard with Header}
\usage{
write_clip_with_header(tbl, desc = NULL, ...)
}
\arguments{
\item{tbl}{Tibble or data frame for writing out to clipboard}

\item{desc}{Optional description}

\item{...}{Optional additional parameters to be passed to internal functions}
}
\value{
Writes table to your clipboard
}
\description{
Writes to Clipboard with Header
}
